Camera and moving target

+++ src/map/position.rs
@@ -2,12 +2,12 @@ use crate::prelude::*;
 
 #[derive(Component, PartialEq)]
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Position {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Position {
             x,
             y
@@ -15,7 +15,7 @@ impl Position {
     }
 
     pub fn zero() -> Self {
-        Position::new(0, 0)
+        Position::new(0.0, 0.0)
     }
 }
